Moved cache config and properties into web.client package
