+++ src/test/java/com/rackspace/salus/monitor_management/web/client/ZoneApiClientTest.java
@@ -23,7 +23,6 @@ import static org.springframework.test.web.client.response.MockRestResponseCreat
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.rackspace.salus.monitor_management.config.ZoneApiCacheConfig;
 import com.rackspace.salus.monitor_management.web.model.ZoneDTO;
 import com.rackspace.salus.telemetry.repositories.TenantMetadataRepository;
 import java.util.List;
